central_service/app/requirements.txt
celery
#psycopg2 # for production
django
djangorestframework
djangorestframework-xml
pytest-django
pytest-xdist
